refactor: :truck: split player classes
